move the printing to conftest such that pytest doesnt ignore it
